Tested BadBroyden on the guess: [-0.9,1.0] on the Rosenbrock.
I managed to get same plots for the good broyden - the algorithm seemed to
find the optimization point but the path there and the plot looked not she slightest least like the first plot so I assumed it was something wrong with my code and discarded it. 

I would suggest you check with these values and observbe the weird plot,
I can accept that it might take a weird path down to the optimization point - but since the whole plot deforms it looks like something is wrong - but maybe it's just the dimensions of the plot that changes and has nothing to do with the correctness od the algorithm.

Anyway - check with that guess on the rosenbrock and verify that Badbroyden actually works - in that case i might know how to fix Good broyden.

git-svn-id: https://xp-dev.com/svn/FMNN25_Projects/trunk@73 af13a1c6-311e-4a12-858b-cde079fe1ede